--- a/python/sglang/srt/layers/activation.py
+++ b/python/sglang/srt/layers/activation.py
@@ -375,9 +375,9 @@ def get_cross_encoder_activation_function(config: PretrainedConfig):
 
 
 if not (
-    _is_cuda or _is_npu or (_is_cpu and _is_cpu_amx_available) or _is_hip or _is_xpu
+    _is_cuda or _is_npu or _is_cpu or _is_hip or _is_xpu
 ):
     logger.info(
-        "sgl-kernel is not available on Non-NV, Non-AMD platforms or Non-AMX CPUs. Fallback to other kernel libraries."
+        "sgl-kernel is not available on Non-NV platforms or CPU runtime. Fallback to other kernel libraries."
     )
     from vllm.model_executor.layers.activation import GeluAndMul, SiluAndMul