--- a/python/sglang/srt/layers/layernorm.py
+++ b/python/sglang/srt/layers/layernorm.py
@@ -329,9 +329,7 @@ class Gemma3RMSNorm(CustomOp):
         return f"{tuple(self.weight.shape)}, eps={self.eps}"
 
 
-if not (
-    _is_cuda or _is_hip or _is_npu or _is_cpu or _is_xpu
-):
+if not (_is_cuda or _is_hip or _is_npu or _is_cpu or _is_xpu):
     logger.info(
         "sgl-kernel layernorm implementation is not available on current platform. Fallback to other kernel libraries."
     )