--- a/python/sglang/srt/layers/layernorm.py
+++ b/python/sglang/srt/layers/layernorm.py
@@ -330,7 +330,7 @@ class Gemma3RMSNorm(CustomOp):
 
 
 if not (
-    _is_cuda or _is_hip or _is_npu or (_is_cpu and _is_cpu_amx_available) or _is_xpu
+    _is_cuda or _is_hip or _is_npu or _is_cpu or _is_xpu
 ):
     logger.info(
         "sgl-kernel layernorm implementation is not available on current platform. Fallback to other kernel libraries."