--- a/python/sglang/srt/layers/quantization/fp8.py
+++ b/python/sglang/srt/layers/quantization/fp8.py
@@ -108,6 +108,9 @@ if _is_hip and (_use_aiter or _use_hip_int4):
     from aiter.fused_moe import fused_moe
     from aiter.ops.shuffle import shuffle_weight
 
+if not (_is_cuda or _is_npu or _is_cpu or _is_hip):
+    from vllm._custom_ops import scaled_fp8_quant
+
 
 ACTIVATION_SCHEMES = ["static", "dynamic"]