@yanbing-j 2025-05-08
Hi @YangQun1, is there any update of this support MLA for torch_native attention backend? Could you please also take CPU into account? Thanks!

cc @mingfeima @chunyuan-w