@yanbing-j 2025-07-18
> I will also test more models on local machines for torch_native backend.

Hi @zhyncs ,

With `--device cpu --attention-backend torch_native`, V2-lite can run now. R1 FP8 and R1 INT8 can also run with  https://github.com/sgl-project/sglang/pull/7820.

Now, CI only has V2-lite for CPU. So this PR is okay to merge.  And we are enpanding model scope to include R1 inside.