@yanbing-j 2025-07-22
Hi @zhyncs , 

I checked https://github.com/sgl-project/sglang/actions/runs/16434788523/job/46443992313?pr=3475, which will fail in mi325 and pass in mi300. I tried to test in local mi300 with cmdline in https://docs.sglang.ai/start/install.html#method-2-from-source, and UT indeed passes in mi300. Could you please give some advices about this failure in mi325? Thanks!

Other CI failuers are not related to the code change of this PR.